[6] fix: Add a conditional compilation alternative to TMPRO
In this commit, we provide a conditional compilation branching in
`ScoreTrackerBehaviour`, an approach we wanted to abandon, however the
Github agent wouldn't have access to TextMeshPro, and therefore we must
provide the alternative compilation using Unity UI's `Text`.

We add a Define constant to the csproj if it can find the TextMeshPro dll,
and that is what we use in the conditional compilation.