[6] fix: Replace the unity path property with unity version
In this commit, we replace the Unity path property in the directory build props at the
root with UnityVersion.
While it won't automatically update the editor version if we ever update, and will not be compatible across people with different versions, it means I can delete all non root props file, but more importantly
hopefully the CI pipeline will work.